delete blurs if image is deleted from media library

+++ src/Plugin.php
@@ -49,7 +49,8 @@ class Plugin {
 
   public function add_hooks(): void {
     add_filter( "wp_generate_attachment_metadata", [ $this, "generate_blur_for_attachment" ], 10, 2);
-    add_filter( 'attachment_fields_to_edit', [ $this, "render_blur_data_in_edit_view" ], 10, 2 );
+    add_filter( "attachment_fields_to_edit", [ $this, "render_blur_data_in_edit_view" ], 10, 2 );
+    add_action( "delete_attachment", [ $this, "remove_blurs_for_removed_attachment" ] );
   }
 
   /**
@@ -144,9 +145,24 @@ class Plugin {
   public function deactivate() {
     $ids = $this->image_repository->get_all_image_ids();
     $sizes = $this->image_repository->get_all_image_sizes_with_default();
-    foreach ($ids as $id) {
+    foreach ( $ids as $id ) {
+      foreach ( $sizes as $size ) {
+        $this->image_blur_repository->delete( $id, $size );
+      }
+    }
+  }
+
+  /**
+   * A function that removes blur data for deleted image. This is added to delete_attachment action.
+   * 
+   * @param int $id - attachment's ID
+   * @return void
+   */
+  public function remove_blurs_for_removed_attachment( int $id ): void {
+    if ( $this->image_repository->is_image( $id ) ) {
+      $sizes = $this->image_repository->get_all_image_sizes_with_default();
       foreach ($sizes as $size) {
-        $this->image_blur_repository->delete($id, $size);
+        $this->image_blur_repository->delete( $id, $size );
       }
     }
   }
